processors/graph/requirements.txt
kafka-python>=2.0.2
gqlalchemy>=1.6.0
asyncpg>=0.29.0
